+++ assets/fonts/stylesheet.css
@@ -1,0 +1,11 @@
+@font-face {
+    font-family: 'liberation_sansregular';
+    src: url('LiberationSans-Regular-webfont.eot');
+    src: url('LiberationSans-Regular-webfont.eot?#iefix') format('embedded-opentype'),
+         url('LiberationSans-Regular-webfont.woff') format('woff'),
+         url('LiberationSans-Regular-webfont.ttf') format('truetype'),
+         url('LiberationSans-Regular-webfont.svg#liberation_sansregular') format('svg');
+    font-weight: normal;
+    font-style: normal;
+}
+
