+++ assets/js/module.ts
@@ -1,0 +1,3 @@
+const add = (a: number, b: number): number => a + b;
+
+export default add;
