+++ assets/scss/_atoms.scss
@@ -1,0 +1,2 @@
+@import "atoms/colors";
+@import "atoms/typography";
