+++ assets/scss/style.scss
@@ -1,0 +1,52 @@
+//@import "../../node_modules/normalize.css/_normalize.css";
+@import "../../node_modules/breakpoint-sass/stylesheets/breakpoint";
+@import "atoms";
+@import "molecules";
+@import "organisms";
+
+@font-face {
+    font-family: 'LiberationSans';
+    src: url('/fonts/LiberationSans-Regular.woff') format('woff'),
+         url('/fonts/LiberationSans-Regular.ttf')  format('truetype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+*,
+*:before,
+*:after {
+    box-sizing: border-box;
+}
+
+html,
+body {
+    margin: 0;
+    padding: 0;
+}
+
+html {
+    font-size: 1em;
+}
+
+body {
+    background-color: $grey-800;
+    color: $grey-500;
+    font-family: 'LiberationSans', Arial, Helvetica, sans-serif;
+    font-weight: 400;
+    line-height: 1.45;
+}
+
+.main {
+    margin: 0;
+    padding: 0;
+}
+
+.section-container {
+    padding: 27px 29px;
+    width: 100%;
+
+    @include breakpoint(min-width 1024px) {
+        padding: 33px 180px;
+        max-width: 100%;
+    }
+}
