--- src/main/resources/static/js/index.ts
@@ -1,6 +1,0 @@
-import add from 'module';
-
-const a: number = 2;
-const b: number = 1;
-
-console.log(add(1, 2));
