--- src/main/resources/static/js/module.ts
@@ -1,3 +1,0 @@
-const add = (a: number, b: number): number => a + b;
-
-export default add;
