--- src/main/resources/static/scss/style.scss
@@ -1,3 +1,0 @@
-body {
-  background-color: red;
-}
